refactor testing functionality into callbacks
