+++ amplfi/train/models/__init__.py
@@ -1,2 +1,3 @@
+from .base import AmplfiModel
 from .flow import FlowModel
 from .similarity import SimilarityModel
